fix tests

+++ internal/signer/local_signer/legacy.go
@@ -47,7 +47,7 @@ func legacyDeriveKeyset(mintKey *bip32.Key, seed cashu.Seed) ([]cashu.MintKey, e
 		return nil, fmt.Errorf("mintKey.NewChildKey(uint32(unit.EnumIndex())). %w", err)
 	}
 
-	versionKey, err := unitKey.NewChildKey(uint32(seed.Version))
+	versionKey, err := unitKey.NewChildKey(seed.Version)
 	if err != nil {
 		return nil, fmt.Errorf("mintKey.NewChildKey(uint32(seed.Version)) %w", err)
 	}
